+++ frontend/src/components/NavigationBar/NavigationBar.tsx
@@ -3,7 +3,7 @@ import * as S from './NavigationBar.style';
 import { useLocation, useNavigate } from 'react-router-dom';
 
 import GET_ROUTES from '@_common/getRoutes';
-import NavigationBarItem from '@_components/NavigationBarItem/NavigationBarItem';
+import NavigationBarItem from '@_components/NavigationBar/NavigationBarItem/NavigationBarItem';
 import { useState } from 'react';
 import { useTheme } from '@emotion/react';
