+++ frontend/src/components/NavigationBar/NavigationBarItem/NavigationBarItem.style.ts
