+++ frontend/src/pages/Moim/MainPage/components/HomeMainContent/MoimCardList/MoimCard/MoimCard.stories.tsx
