+++ frontend/src/pages/Moim/MainPage/components/HomeMainContent/MoimList/MoimCardListSkeleton/MoimCardListSkeleton.tsx
