+++ frontend/src/pages/Moim/MainPage/components/HomeMainContent/MoimList/MoimCardListSkeleton/MoimCardSkeleton/MoimCardSkeleton.stories.tsx
