+++ frontend/src/pages/Moim/MainPage/components/MoimTabBar/MoimTabBar.stories.tsx
