+++ frontend/src/pages/Moim/MoimDetailPage/components/CommentList/CommentList.tsx
@@ -1,11 +1,11 @@
-import * as S from '@_components/CommentList/ComentList.style';
+import * as S from './ComentList.style';
 
 import { HTMLProps, useState } from 'react';
 
 import { Comment } from '@_types/index';
-import CommentCard from '@_components/CommentList/CommentCard/CommentCard';
 import MessageInput from '@_components/Input/MessagInput/MessageInput';
 import useWriteComment from '@_hooks/mutaions/useWriteComment';
+import CommentCard from './CommentCard/CommentCard';
 
 interface CommentListProps extends HTMLProps<HTMLDivElement> {
   moimId: number;
