+++ frontend/src/pages/Moim/MoimDetailPage/components/MoimInformation/MoimInfomationSkeleton.tsx
