+++ frontend/src/pages/Moim/MoimDetailPage/components/MoimInformation/MoimInformation.style.ts
