+++ frontend/src/pages/Mypage/components/MineInfoCard/MineInfoCard.style.ts
