+++ frontend/src/pages/Mypage/components/MineInfoCard/MoinInfoCard.stories.tsx
